+++ src/managedcleanroom/HISTORY.rst
@@ -6,3 +6,8 @@ Release History
 1.0.0b1
 ++++++
 * Initial release.
+
+1.0.0b2
+++++++
+* Add frontend commandlets
+* Add MSAL device code flow authentication